Serialization project and scene

+++ Nuake/src/Core/Physics/BulletDebugDrawer.cpp
@@ -11,9 +11,9 @@ void BulletDebugDrawer::drawLine(const btVector3& from, const btVector3& to, con
 	tc.Rotation = glm::vec3(0.0f);
 
 	glBegin(GL_LINES);
-	glColor3f(color.x(), color.y(), color.z());
-	glVertex3f(from.x(), from.y(), from.z());
-	glVertex3f(to.x(), to.y(), to.z());
+	glColor3f((float)(color.x()), (float)(color.y()), (float)(color.z()));
+	glVertex3f((float)(from.x()), (float)(from.y()), (float)(from.z()));
+	glVertex3f((float)(to.x()), (float)(to.y()), (float)(to.z()));
 	glEnd();
 
 	//Renderer::DrawCube(tc, glm::vec4(color.x(), color.y(), color.z(), 1.0f));
